Consolidate Pike db modules
This consolidates the agents and organizations db modules into one
"pike" module. The functionality doesn't change at all but now all the
pike operations are in one place. This simplifies some code and also
makes it clearer that these modules are for the Pike smart contract.
