+++ cli/src/actions/agents.rs
@@ -19,7 +19,7 @@ use crate::error::CliError;
 use crate::http::submit_batches;
 use crate::transaction::pike_batch_builder;
 use grid_sdk::{
-    agents::addressing::PIKE_NAMESPACE,
+    pike::addressing::PIKE_NAMESPACE,
     protocol::pike::payload::{Action, CreateAgentAction, PikePayloadBuilder, UpdateAgentAction},
     protos::IntoProto,
 };
