+++ cli/src/actions/organizations.rs
@@ -19,7 +19,7 @@ use crate::error::CliError;
 use crate::http::submit_batches;
 use crate::transaction::pike_batch_builder;
 use grid_sdk::{
-    organizations::addressing::PIKE_NAMESPACE,
+    pike::addressing::PIKE_NAMESPACE,
     protocol::pike::payload::{
         Action, CreateOrganizationAction, PikePayloadBuilder, UpdateOrganizationAction,
     },
