+++ contracts/location/src/handler.rs
@@ -410,8 +410,7 @@ mod tests {
     use std::collections::HashMap;
 
     use grid_sdk::{
-        agents::addressing::compute_agent_address,
-        organizations::addressing::compute_organization_address,
+        pike::addressing::{compute_agent_address, compute_organization_address},
         protocol::{
             location::payload::{
                 LocationCreateActionBuilder, LocationDeleteActionBuilder,
