+++ contracts/pike/src/handler.rs
@@ -30,9 +30,8 @@ cfg_if! {
 }
 
 use grid_sdk::{
-    agents::addressing::{compute_agent_address, PIKE_NAMESPACE},
-    organizations::addressing::compute_organization_address,
     permissions::PermissionChecker,
+    pike::addressing::{compute_agent_address, compute_organization_address, PIKE_NAMESPACE},
     protos::{
         pike_payload::{
             CreateAgentAction, CreateOrganizationAction, PikePayload, PikePayload_Action as Action,
