+++ contracts/track_and_trace/src/state.rs
@@ -23,7 +23,7 @@ cfg_if! {
 }
 
 use grid_sdk::{
-    agents::addressing::compute_agent_address,
+    pike::addressing::compute_agent_address,
     protocol::{
         pike::state::{Agent, AgentList},
         schema::state::{Schema, SchemaList},
