+++ sdk/src/pike/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2018-2020 Cargill Incorporated
+// Copyright 2018-2021 Cargill Incorporated
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,26 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-table! {
-    agent (id) {
-        id -> Int8,
-        public_key -> Varchar,
-        org_id -> Varchar,
-        active -> Bool,
-        metadata -> Binary,
-        start_commit_num -> Int8,
-        end_commit_num -> Int8,
-        service_id -> Nullable<Text>,
-    }
-}
+pub mod addressing;
+pub mod store;
 
-table! {
-    role (id) {
-        id -> Int8,
-        public_key -> Varchar,
-        role_name -> Text,
-        start_commit_num -> Int8,
-        end_commit_num -> Int8,
-        service_id -> Nullable<Text>,
-    }
-}
+#[cfg(feature = "diesel")]
+pub use store::diesel::DieselPikeStore;
+pub use store::PikeStore;
