+++ sdk/src/pike/store/diesel/operations/list_organizations.rs
@@ -12,37 +12,33 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::OrganizationStoreOperations;
+use super::PikeStoreOperations;
 use crate::commits::MAX_COMMIT_NUM;
 use crate::error::InternalError;
-use crate::organizations::store::diesel::models::OrganizationModel;
-use crate::organizations::store::diesel::{schema::organization, OrganizationStoreError};
-use crate::{
-    organizations::store::{Organization, OrganizationList},
-    paging::Paging,
-};
+use crate::paging::Paging;
+use crate::pike::store::diesel::models::OrganizationModel;
+use crate::pike::store::diesel::{schema::organization, PikeStoreError};
+use crate::pike::store::{Organization, OrganizationList};
 
 use diesel::prelude::*;
 
-pub(in crate::organizations::store::diesel) trait OrganizationStoreListOrganizationsOperation {
+pub(in crate::pike::store::diesel) trait PikeStoreListOrganizationsOperation {
     fn list_organizations(
         &self,
         service_id: Option<&str>,
         offset: i64,
         limit: i64,
-    ) -> Result<OrganizationList, OrganizationStoreError>;
+    ) -> Result<OrganizationList, PikeStoreError>;
 }
 
 #[cfg(feature = "postgres")]
-impl<'a> OrganizationStoreListOrganizationsOperation
-    for OrganizationStoreOperations<'a, diesel::pg::PgConnection>
-{
+impl<'a> PikeStoreListOrganizationsOperation for PikeStoreOperations<'a, diesel::pg::PgConnection> {
     fn list_organizations(
         &self,
         service_id: Option<&str>,
         offset: i64,
         limit: i64,
-    ) -> Result<OrganizationList, OrganizationStoreError> {
+    ) -> Result<OrganizationList, PikeStoreError> {
         let mut query = organization::table
             .into_boxed()
             .select(organization::all_columns)
@@ -59,7 +55,7 @@ impl<'a> OrganizationStoreListOrganizationsOperation
         let orgs = query
             .load::<OrganizationModel>(self.conn)
             .map_err(|err| {
-                OrganizationStoreError::InternalError(InternalError::from_source(Box::new(err)))
+                PikeStoreError::InternalError(InternalError::from_source(Box::new(err)))
             })?
             .into_iter()
             .map(Organization::from)
@@ -85,15 +81,15 @@ impl<'a> OrganizationStoreListOrganizationsOperation
 }
 
 #[cfg(feature = "sqlite")]
-impl<'a> OrganizationStoreListOrganizationsOperation
-    for OrganizationStoreOperations<'a, diesel::sqlite::SqliteConnection>
+impl<'a> PikeStoreListOrganizationsOperation
+    for PikeStoreOperations<'a, diesel::sqlite::SqliteConnection>
 {
     fn list_organizations(
         &self,
         service_id: Option<&str>,
         offset: i64,
         limit: i64,
-    ) -> Result<OrganizationList, OrganizationStoreError> {
+    ) -> Result<OrganizationList, PikeStoreError> {
         let mut query = organization::table
             .into_boxed()
             .select(organization::all_columns)
@@ -110,7 +106,7 @@ impl<'a> OrganizationStoreListOrganizationsOperation
         let orgs = query
             .load::<OrganizationModel>(self.conn)
             .map_err(|err| {
-                OrganizationStoreError::InternalError(InternalError::from_source(Box::new(err)))
+                PikeStoreError::InternalError(InternalError::from_source(Box::new(err)))
             })?
             .into_iter()
             .map(Organization::from)
